Refactor package structure in gateway
The package structure of the gateway module has been significantly refactored to better reflect its purpose. This largely involves renaming the 'event_sourcing' part of the package to 'messaging.graphql', to better represent that this module is now primarily used for GraphQL messaging.
