+++ messaging/graphql/gateway/src/main/java/nawaphon/microservices/messaging/graphql/gateway/components/FakeDatabaseComponent.java
@@ -1,9 +1,9 @@
-package nawaphon.microservices.event_sourcing.gateway.components;
+package nawaphon.microservices.messaging.graphql.gateway.components;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nawaphon.microservices.event_sourcing.gateway.pojo.Customer;
-import nawaphon.microservices.event_sourcing.gateway.pojo.CustomerDetail;
+import nawaphon.microservices.messaging.graphql.gateway.pojo.Customer;
+import nawaphon.microservices.messaging.graphql.gateway.pojo.CustomerDetail;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Component;
