+++ messaging/graphql/gateway/src/main/java/nawaphon/microservices/messaging/graphql/gateway/pojo/CustomerDetail.java
@@ -1,4 +1,4 @@
-package nawaphon.microservices.event_sourcing.gateway.pojo;
+package nawaphon.microservices.messaging.graphql.gateway.pojo;
 
 import java.util.UUID;
